slt/requirements.txt
absl-py==0.9.0
asn1crypto==1.2.0
astor==0.8.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
# conda==4.8.0
# conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
gast==0.2.2
google-auth==1.10.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyboard==0.13.5
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
more-itertools==8.2.0
natsort==7.0.1
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
packaging==20.3
pandas==1.0.3
Pillow==7.1.0
plotly==4.5.0
pluggy==0.13.1
portalocker==1.5.2
protobuf==3.11.2
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
pytest==5.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
ruamel-yaml==0.15.87
sacrebleu==1.4.4
scikit-learn==0.22.2.post1
scipy==1.4.1
sentencepiece==0.1.85
six==1.13.0
# tensorboard==2.1.2
tensorflow==2.1.2
# tensorflow-estimator==2.1.2
termcolor==1.1.0
torch==1.4.0
torchtext==0.5.0
torchvision==0.5.0
tqdm==4.40.2
typing==3.7.4.1
urllib3==1.25.7
# warmup-scheduler==0.1.1
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
xmltodict==0.12.0
zipp==3.1.0